+++ modules/output.js
@@ -15,7 +15,8 @@ var warn = function(){
 exports.cmd = function(input, encoder){
   try {
     ncp.copy(input, function () {
-        console.log(log.yellow(input));
+        process.stdout.write(input);
+
         if (encoder){
           process.exit(1);
         }